test/fixtures/wpt/resource-timing/iframe-non-html.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Test the sequence of events when reporting iframe timing.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/resource-loaders.js"></script>
<body>
<script>
    function test(href, type) {
        promise_test(async t => {
            await load.iframe(href);
            const entries = performance.getEntriesByType('resource').filter(({name}) => name.includes(href));
            assert_equals(entries.length, 1);
            assert_equals(entries[0].initiatorType, 'iframe');
        }, `Iframes should report resource timing for ${type} iframes`);
    }

    test('/common/square.png', 'image');
    test('/common/dummy.xhtml', 'xhtml');
    test('/common/dummy.xml', 'xml');
    test('/common/text-plain.txt', 'text');
</script>
</body>
